migrate test_google_chat to pytest
